% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llama.R
\name{llama_infer}
\alias{llama_infer}
\title{Infer from a llama model}
\usage{
llama_infer(prompt = "What's the capital of Spain?", model_path = NULL)
}
\arguments{
\item{prompt}{The prompt to infer from}

\item{model_path}{The path to the model}
}
\value{
The response from the model
}
\description{
Infer from a llama model
}
